update notes routes and controller using authentication middleware

+++ src/controllers/NotesController.js
@@ -1,9 +1,10 @@
 const knex = require("../database/knex");
+const AppError = require("../utils/AppError");
 
 class NotesController {
   async create(req, res) {
     const { title, description, rating, tags } = req.body;
-    const { user_id } = req.params;
+    const user_id = req.user.id;
 
     const movie_id = await knex("movie_notes").insert({
       title,
@@ -29,14 +30,24 @@ class NotesController {
 
   async show(req, res) {
     const { id } = req.params;
+    const user_id = req.user.id;
 
     const movie = await knex("movie_notes").where({ id }).first();
+    if (movie.user_id !== user_id) {
+      throw new AppError("Você não tem permissão para acessar essa nota", 401);
+    }
 
     return res.json(movie);
   }
 
   async delete(req, res) {
     const { id } = req.params;
+    const user_id = req.user.id;
+
+    const movie = await knex("movie_notes").where({ id }).first();
+    if (movie.user_id !== user_id) {
+      throw new AppError("Você não tem permissão para excluir essa nota", 401);
+    }
 
     await knex("movie_notes").where({ id }).first().delete();
 
@@ -44,11 +55,11 @@ class NotesController {
   }
 
   async showAll(req, res) {
-    const { user_id } = req.query;
+    const user_id = req.user.id;
 
-    const notes = await knex("movie_notes").where({ user_id });
+    const notes = await knex("movie_notes").where({ user_id }).orderBy("created_at", "desc");
 
-    return res.json(notes)
+    return res.json({ notes })
   }
 }
